Velocity Swap and Attempted (Failed) TraceParams
Momentum is NOT being swapped, because I haven't been able to figure out how to access mass. Currently raw velocity is what is swapped.

I tried to do a traceparam by doing a multitrace, and setting the floor to ignore visibiliity to traces, but that seemed to have no effect. Could be that the raytrace still stores it in the Tarray, and I have to still move past it in some form? idk.